rename manifest_marker -> manifest_maker
