+++ src/python/ensembl/io/genomio/manifest_maker.py
